+++ client/src/components/PropertiesSummary.tsx
@@ -17,6 +17,7 @@ import { I18n } from "@lingui/core/i18n";
 import { I18n as I18nComponent } from "@lingui/react";
 import { PortfolioGraph } from "./PortfolioGraph";
 import _ from "lodash";
+import helpers from "util/helpers";
 
 type Props = withMachineInStateProps<"portfolioFound"> & {
   isVisible: boolean;
@@ -40,74 +41,6 @@ const generateLinkToDataRequestForm = (i18n: I18n, address?: AddressRecord) => {
     url = `${url}?prefill_Address=${fullAddress}`;
   }
   return url;
-};
-
-const getMostCommonElementsInArray = (array: any[], numberOfResults: number) => {
-  const elementsByFrequency = _.countBy(array);
-  const sortedElementsByFrequency = Object.entries(elementsByFrequency).sort((a, b) => b[1] - a[1]);
-  return sortedElementsByFrequency.slice(0, numberOfResults).map((a) => a[0]);
-};
-
-const getTopFiveContactsInPortfolio = (addrs: AddressRecord[]) => {
-  const allContactNames = addrs
-    .map((a) => a.ownernames?.map((ownername) => ownername.value) || [])
-    .flat();
-
-  return getMostCommonElementsInArray(allContactNames, 5);
-};
-
-export const calculateAggDataFromAddressList = (addrs: AddressRecord[]): SummaryStatsRecord => {
-  const bldgs = addrs.length;
-  const units = _.sumBy(addrs, (a) => a.unitsres || 0);
-  const totalopenviolations = _.sumBy(addrs, (a) => a.openviolations);
-  const totalviolations = _.sumBy(addrs, (a) => a.totalviolations);
-  const totalevictions = _.sumBy(addrs, (a) => a.evictions || 0);
-  const totalrsgain = addrs.reduce(
-    (sum, a) => (a.rsdiff && a.rsdiff > 0 ? sum + a.rsdiff : sum),
-    0
-  );
-  const totalrsloss = addrs.reduce(
-    (sum, a) => (a.rsdiff && a.rsdiff < 0 ? sum + a.rsdiff : sum),
-    0
-  );
-  const totalrsdiff = totalrsgain + totalrsloss;
-
-  const addrWithBiggestRsLoss = addrs.reduce((a1, a2) =>
-    (a1.rsdiff || 0) < (a2.rsdiff || 0) ? a1 : a2
-  );
-
-  const addrWithMostEvictions = addrs.reduce((a1, a2) =>
-    (a1.evictions || 0) > (a2.evictions || 0) ? a1 : a2
-  );
-
-  const addrWithMostOpenViolations = addrs.reduce((a1, a2) =>
-    (a1.openviolations || 0) > (a2.openviolations || 0) ? a1 : a2
-  );
-
-  const allBusinessAddrs = addrs.map((addr) => addr.businessaddrs || []).flat();
-  const allCorpNames = addrs.map((addr) => addr.corpnames || []).flat();
-
-  return {
-    bldgs,
-    units,
-    age: new Date().getFullYear() - _.meanBy(addrs, (a) => a.yearbuilt),
-    topowners: getTopFiveContactsInPortfolio(addrs),
-    topcorp: getMostCommonElementsInArray(allCorpNames, 1)[0] || null,
-    topbusinessaddr: getMostCommonElementsInArray(allBusinessAddrs, 1)[0] || null,
-    totalopenviolations,
-    totalviolations,
-    openviolationsperbldg: totalopenviolations / bldgs,
-    openviolationsperresunit: totalopenviolations / units,
-    totalevictions,
-    avgevictions: totalevictions / bldgs,
-    totalrsgain,
-    totalrsloss,
-    totalrsdiff,
-    rsproportion: (Math.abs(totalrsdiff) / units) * 100,
-    rslossaddr: { ...addrWithBiggestRsLoss },
-    evictionsaddr: { ...addrWithMostEvictions },
-    violationsaddr: { ...addrWithMostOpenViolations },
-  };
 };
 
 const DataRequestButton: React.FC<{
